Fix build for 2022.1 EAP
The refreshVersions plugin uses the Kotlin Gradle plugin. The problem here is that the first time a Gradle plugin is activated determines the used version, which for the Kotlin plugin also automatically determines the Kotlin language version. refreshVersions uses 1.4.20, which is incompatible with the Intellij 2022.1 sources, which were compiled using Kotlin 1.6.0. Adding the Kotlin plugin in the correct version in settings.gradle.kts before refreshVersions without applying it fixes its version to 1.6.0 for the whole project and makes everything compile as it should.